hw2/condaenv.a92jjrsn.requirements.txt
cycler==0.10.0
future==0.18.2
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pillow==6.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyyaml==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
torch==1.5.0
torchvision==0.6.0
yacs==0.1.7